+++ src/main/java/at/jku/dke/etutor/task_app/test/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains classes for a simple load test.
+ */
+package at.jku.dke.etutor.task_app.test;